Initial memoize implementation
This uses moize because it allows configuring the cache size, unlike
fast-memoize.

However, it swells the content script from 66610 bytes to 80964 bytes
which seems excessive.

We should just adapt fast-memoize to do the cache size check (and add
Promise support) and pare back some of the stuff we don't need from it
too.